requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
alphashape==1.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
beautifulsoup4==4.14.3
brotli==1.2.0
certifi==2026.2.25
cffi==2.0.0
chardet==7.0.1
charset-normalizer==3.4.5
click==8.3.1
click-log==0.4.0
colorama==0.4.6
cryptography==46.0.5
cssselect==1.4.0
datasets==4.6.1
dill==0.4.0
distro==1.9.0
fake-http-header==0.3.5
fake-useragent==2.2.0
fastuuid==0.14.0
filelock==3.25.0
frozenlist==1.8.0
fsspec==2026.2.0
greenlet==3.3.2
h11==0.16.0
h2==4.3.0
hf-xet==1.3.2
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.0
humanize==4.15.0
hyperframe==6.1.0
idna==3.11
importlib-metadata==8.7.1
jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lark==1.3.1
litellm==1.82.0
lxml==5.4.0
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
multidict==6.7.1
multiprocess==0.70.18
networkx==3.6.1
nltk==3.9.3
numpy==2.4.2
openai==2.26.0
packaging==26.0
pandas==3.0.1
patchright==1.58.1
pillow==12.1.1
# crawl4ai and Playwright removed: using aiohttp + Trafilatura instead
propcache==0.4.1
psutil==7.2.2
pyarrow==23.0.1
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.13.1
pyee==13.0.1
pygments==2.19.2
pyopenssl==25.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-telegram-bot==22.6
pyyaml==6.0.3
rank-bm25==0.2.2
rapidfuzz==3.14.5
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
rtree==1.4.1
scipy==1.17.1
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.8.3
# tf-playwright-stealth removed
tiktoken==0.12.0
tokenizers==0.22.2
transformers==4.56.2
tqdm==4.67.3
trafilatura==2.0.0
trimesh==4.11.3
typer==0.24.1
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
urllib3==2.6.3
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
